Move `dummy_segmentation` into `test_process_seg.py`
Notes:

- `dummy_segmentation` is used *only* in `test_process_seg.py`; its
  implementation details are specifically tied to those tests.
- `dummy_segmentation_4d` was never used at all.
